split sync functions to new file
