+++ utilsChecker.py
@@ -13,6 +13,7 @@ import shutil
 import utilsDataman
 import requests
 import ffmpeg
+import logging
 import matplotlib.pyplot as plt
 from scipy.ndimage import gaussian_filter1d
 from scipy.signal import gaussian, sosfiltfilt, butter, find_peaks
@@ -905,382 +906,6 @@ def synchronizeVideos(CameraDirectories, trialRelativePath, pathPoseDetector,
         
     return pointDir, confDir, markerNames, frameRate, nansInOutDir, startEndFrames, cameras2Use
 
-# %%
-def synchronizeVideoKeypoints(keypointList, confidenceList,
-                              confidenceThreshold=0.3, 
-                              filtFreqs = {'gait':12,'default':500},
-                              sampleFreq=30, visualize=False, maxShiftSteps=30,
-                              isGait=False, CameraParams = None,
-                              cameras2Use=['none'],CameraDirectories = None,
-                              trialName=None, trialID=''):
-    visualize2Dkeypoint = False # this is a visualization just for testing what filtered input data looks like
-    
-    # keypointList is a mCamera length list of (nmkrs,nTimesteps,2) arrays of camera 2D keypoints
-    print('Synchronizing Keypoints')
-    
-    # Deep copies such that the inputs do not get modified.
-    c_CameraParams = copy.deepcopy(CameraParams)
-    c_cameras2Use = copy.deepcopy(cameras2Use)
-    c_CameraDirectoryDict = copy.deepcopy(CameraDirectories)
-    
-    # Turn Camera Dict into List
-    c_CameraDirectories = list(c_CameraDirectoryDict.values())
-    # Check if one camera has only 0s as confidence scores, which would mean
-    # no one has been properly identified. We want to kick out this camera
-    # from the synchronization and triangulation. We do that by popping out
-    # the corresponding data before syncing and add back 0s later.
-    badCameras = []
-    for icam, conf in enumerate(confidenceList):
-        if np.max(conf) == 0.0:
-            badCameras.append(icam)
-    idxBadCameras = [badCameras[i]-i for i in range(len(badCameras))]
-    cameras2NotUse = []
-    for idxBadCamera in idxBadCameras:
-        print('{} kicked out of synchronization'.format(
-            c_cameras2Use[idxBadCamera]))
-        cameras2NotUse.append(c_cameras2Use[idxBadCamera])
-        keypointList.pop(idxBadCamera)
-        confidenceList.pop(idxBadCamera)
-        c_CameraParams.pop(idxBadCamera)
-        c_cameras2Use.pop(idxBadCamera)
-        c_CameraDirectories.pop(idxBadCamera)
-        
-        
-    markerNames = getOpenPoseMarkerNames()
-    mkrDict = {mkr:iMkr for iMkr,mkr in enumerate(markerNames)}
-    
-    # First, remove occluded markers
-    footMkrs = {'right':[mkrDict['RBigToe'], mkrDict['RSmallToe'], mkrDict['RHeel'],mkrDict['RAnkle']],
-                'left':[mkrDict['LBigToe'], mkrDict['LSmallToe'], mkrDict['LHeel'],mkrDict['LAnkle']]}
-    armMkrs = {'right':[mkrDict['RElbow'], mkrDict['RWrist']],
-                'left':[mkrDict['LElbow'], mkrDict['LWrist']]}
-    
-    plt.close('all')
-    
-    # Copy for visualization
-    keypointListUnfilt = copy.deepcopy(keypointList)
-    
-    # remove occluded foot markers (uses large differences in confidence)
-    keypointList,confidenceList = zip(*[removeOccludedSide(keys,conf,footMkrs,confidenceThreshold,visualize=False) for keys,conf in zip(keypointList,confidenceList)])
-    # remove occluded arm markers
-    keypointList,confidenceList = zip(*[removeOccludedSide(keys,conf,armMkrs,confidenceThreshold,visualize=False) for keys,conf in zip(keypointList,confidenceList)])
-       
-    # Copy for visualization 
-    keypointListOcclusionRemoved = copy.deepcopy(keypointList)
-    
-    # Don't change these. The ankle markers are used for gait detector
-    markers4VertVel = [mkrDict['RAnkle'], mkrDict['LAnkle']] # R&L Ankles and Heels did best. There are some issues though - like when foot marker velocity is aligned with camera ray
-    markers4HandPunch = [mkrDict['RWrist'], mkrDict['LWrist'],mkrDict['RShoulder'],mkrDict['LShoulder']]
-    markers4Ankles = [mkrDict['RAnkle'],mkrDict['LAnkle']]
-    
-    # find velocity signals for synchronization
-    nCams = len(keypointList)
-    vertVelList = []
-    mkrSpeedList = []
-    handPunchVertPositionList = []
-    allMarkerList = []
-    for (keyRaw,conf) in zip(keypointList,confidenceList):
-        keyRaw_clean, _, _, _ = clean2Dkeypoints(keyRaw,conf,confidenceThreshold=0.3,nCams=nCams,linearInterp=True)        
-        keyRaw_clean_smooth = smoothKeypoints(keyRaw_clean, sdKernel=3) 
-        handPunchVertPositionList.append(getPositions(keyRaw_clean_smooth,markers4HandPunch,direction=1)) 
-        vertVelList.append(getVertVelocity(keyRaw_clean_smooth)) # doing it again b/c these settings work well for synchronization
-        mkrSpeedList.append(getMarkerSpeed(keyRaw_clean_smooth,markers4VertVel,confidence=conf,averageVels=False)) # doing it again b/c these settings work well for synchronization
-        allMarkerList.append(keyRaw_clean_smooth)
-        
-    # Find indices with high confidence that overlap between cameras.    
-    # Note: Could get creative and do camera pair syncing in the future, based
-    # on cameras with greatest amount of overlapping confidence.
-    overlapInds_clean, minConfLength_all = findOverlap(confidenceList,
-                                                   markers4VertVel)
-    
-    # If no overlap found, try with fewer cameras.
-    c_nCams = len(confidenceList)
-    while not np.any(overlapInds_clean) and c_nCams>2:
-        print("Could not find overlap with {} cameras - trying with {} cameras".format(c_nCams, c_nCams-1))
-        cam_list = [i for i in range(nCams)]
-        # All possible combination with c_nCams-1 cameras.
-        from itertools import combinations
-        combs = set(combinations(cam_list, c_nCams-1))
-        overlapInds_clean_combs = []
-        for comb in combs:
-            confidenceList_sel = [confidenceList[i] for i in list(comb)]
-            overlapInds_clean_c, minConfLength_c = findOverlap(
-                confidenceList_sel, markers4VertVel)
-            overlapInds_clean_combs.append(overlapInds_clean_c.flatten())
-        longest_stretch = 0
-        for overlapInds_clean_comb in overlapInds_clean_combs:
-            stretch_size = overlapInds_clean_comb.shape[0]
-            if stretch_size > longest_stretch:
-                longest_stretch = stretch_size
-                overlapInds_clean = overlapInds_clean_comb
-        c_nCams -= 1
-        
-    # If no overlap found, return 0s.
-    if not np.any(overlapInds_clean):
-        keypointsSync = []
-        confidenceSync = []
-        nansInOutSync = []
-        for i in range(len(cameras2Use)):
-            keypointsSync.insert(i, np.zeros((keypointList[0].shape[0], 10,
-                                              keypointList[0].shape[2])))
-            confidenceSync.insert(i, np.zeros((keypointList[0].shape[0], 10)))
-            nansInOutSync.insert(i, np.array([np.nan, np.nan]))     
-        return keypointsSync, confidenceSync, nansInOutSync
-                
-    [idxStart, idxEnd] = [np.min(overlapInds_clean), np.max(overlapInds_clean)]
-    idxEnd += 1 # Python indexing system.
-    # Take max shift between cameras into account.
-    idxStart = int(np.max([0,idxStart - maxShiftSteps]))
-    idxEnd = int(np.min([idxEnd+maxShiftSteps,minConfLength_all]))
-    # Re-sample the lists    
-    vertVelList = [v[idxStart:idxEnd] for v in vertVelList]
-    mkrSpeedList = [v[:,idxStart:idxEnd] for v in mkrSpeedList]
-    handPunchVertPositionList = [p[:,idxStart:idxEnd] for p in handPunchVertPositionList]
-    allMarkerList = [p[:,idxStart:idxEnd] for p in allMarkerList]
-    confSyncList= [c[:,idxStart:idxEnd] for c in confidenceList]
-    
-    # We do this again, since it might have changed after finding the overlap period.
-    keypointList = list(keypointList)
-    confidenceList = list(confidenceList)
-    badCamerasOverlap = []
-    for icam, conf in enumerate(confSyncList):
-        if np.mean(conf) <= 0.01: # Looser than sum=0 to disregard very few frames with data
-            badCamerasOverlap.append(icam)
-    idxbadCamerasOverlap = [badCamerasOverlap[i]-i for i in range(len(badCamerasOverlap))]
-    for idxbadCameraOverlap in idxbadCamerasOverlap:
-        print('{} kicked out of synchronization - after overlap check'.format(
-            c_cameras2Use[idxbadCameraOverlap]))
-        cameras2NotUse.append(c_cameras2Use[idxbadCameraOverlap])
-        keypointList.pop(idxbadCameraOverlap)
-        confidenceList.pop(idxbadCameraOverlap)
-        c_CameraParams.pop(idxbadCameraOverlap)
-        c_cameras2Use.pop(idxbadCameraOverlap)
-        c_CameraDirectories.pop(idxbadCameraOverlap)
-        
-        vertVelList.pop(idxbadCameraOverlap)
-        mkrSpeedList.pop(idxbadCameraOverlap)
-        handPunchVertPositionList.pop(idxbadCameraOverlap)
-        allMarkerList.pop(idxbadCameraOverlap)
-        confSyncList.pop(idxbadCameraOverlap)        
-    nCams = len(keypointList)
-    
-    # Detect whether it is a gait trial, which determines what sync algorithm
-    # to use. Input right and left ankle marker speeds. Gait should be
-    # detected for all cameras (all but one camera is > 2 cameras) for the
-    # trial to be considered a gait trial.
-    try:
-        isGait = detectGaitAllVideos(mkrSpeedList,allMarkerList,confSyncList,markers4Ankles,sampleFreq)
-    except:
-        isGait = False
-        print('Detect gait activity algorithm failed.')
-    
-    # Detect activity, which determines sync function that gets used
-    isHandPunch,handForPunch = detectHandPunchAllVideos(handPunchVertPositionList,sampleFreq)
-    if isHandPunch:
-        syncActivity = 'handPunch'
-    elif isGait:
-        syncActivity = 'gait'
-    else:
-        syncActivity = 'general'
-        
-    print('Using ' + syncActivity + ' sync function.')
-    
-    
-    # Select filtering frequency based on if it is gait
-    if isGait: 
-        filtFreq = filtFreqs['gait']
-    else: 
-        filtFreq = filtFreqs['default']
-    
-    # Filter keypoint data
-    # sdKernel = sampleFreq/(2*np.pi*filtFreq) # not currently used, but in case using gaussian smoother (smoothKeypoints function) instead of butterworth to filter keypoints
-    keyFiltList = []
-    confFiltList = []
-    confSyncFiltList = []
-    nansInOutList = []
-    for (keyRaw,conf) in zip(keypointList,confidenceList):
-        keyRaw_clean, conf_clean, nans_in_out, conf_sync_clean = clean2Dkeypoints(keyRaw,conf,confidenceThreshold,nCams=nCams)
-        keyRaw_clean_filt = filterKeypointsButterworth(keyRaw_clean,filtFreq,sampleFreq,order=4)
-        keyFiltList.append(keyRaw_clean_filt)
-        confFiltList.append(conf_clean)
-        confSyncFiltList.append(conf_sync_clean)
-        nansInOutList.append(nans_in_out)
-
-    # Copy for visualization
-    keypointListFilt = copy.deepcopy(keyFiltList)
-    confidenceListFilt = copy.deepcopy(confFiltList)
-    confidenceSyncListFilt = copy.deepcopy(confSyncFiltList)
-
-    # find nSample shift relative to the first camera
-    # nSamps = keypointList[0].shape[1]
-    shiftVals = []
-    shiftVals.append(0)
-    timeVecs = []
-    tStartEndVec = np.zeros((len(keypointList),2))
-    for iCam,vertVel in enumerate(vertVelList):
-        timeVecs.append(np.arange(keypointList[iCam].shape[1]))
-        if iCam>0:
-            # if no keypoints in Cam0 or the camera of interest, do not use cross_corr to sync.
-            if np.max(np.abs(vertVelList[iCam])) == 0 or np.max(np.abs(vertVelList[0])) == 0:
-                lag = 0
-            elif syncActivity == 'general':
-                dataForReproj = {'CamParamList':c_CameraParams,
-                                 'keypointList':keypointListFilt,
-                                 'cams2UseReproj': [0, c_cameras2Use.index(c_cameras2Use[iCam])],
-                                 'confidence': confidenceSyncListFilt,
-                                 'cameras2Use': c_cameras2Use
-                                 }
-                corVal,lag = cross_corr(vertVel,vertVelList[0],multCorrGaussianStd=maxShiftSteps/2,
-                                        visualize=False,dataForReproj=dataForReproj,
-                                        frameRate=sampleFreq) # gaussian curve gets multipled by correlation plot - helping choose the smallest shift value for periodic motions
-            elif syncActivity == 'gait':
-                
-                dataForReproj = {'CamParamList':c_CameraParams,
-                                 'keypointList':keypointListFilt,
-                                 'cams2UseReproj': [0, c_cameras2Use.index(c_cameras2Use[iCam])],
-                                 'confidence': confidenceSyncListFilt,
-                                 'cameras2Use': c_cameras2Use
-                                 }
-                corVal,lag = cross_corr_multiple_timeseries(mkrSpeedList[iCam],
-                                            mkrSpeedList[0],
-                                            multCorrGaussianStd=maxShiftSteps/2,
-                                            dataForReproj=dataForReproj,
-                                            visualize=False,
-                                            frameRate=sampleFreq)    
-            elif syncActivity == 'handPunch':
-                corVal,lag = syncHandPunch([handPunchVertPositionList[i] for i in [0,iCam]],
-                                           handForPunch,maxShiftSteps=maxShiftSteps)
-            if np.abs(lag) > maxShiftSteps: # if this fails and we get a lag greater than maxShiftSteps (units=timesteps)
-                lag = 0 
-                print('Did not use cross correlation to sync {} - computed shift was greater than specified {} frames. Shift set to 0.'.format(c_cameras2Use[iCam], maxShiftSteps))
-            shiftVals.append(lag)
-            timeVecs[iCam] = timeVecs[iCam] - shiftVals[iCam]
-        tStartEndVec[iCam,:] = [timeVecs[iCam][0], timeVecs[iCam][-1]]
-        
-    # align signals - will start at the latest-starting frame (most negative shift) and end at
-    # nFrames - the end of the earliest starting frame (nFrames - max shift)
-    tStart = np.max(tStartEndVec[:,0])
-    tEnd = np.min(tStartEndVec[:,1])
-    
-    keypointsSync = []
-    confidenceSync = []
-    startEndFrames = []
-    nansInOutSync = []
-    for iCam,key in enumerate(keyFiltList):
-        # Trim the keypoints and confidence lists
-        confidence = confFiltList[iCam]
-        iStart = int(np.argwhere(timeVecs[iCam]==tStart))
-        iEnd = int(np.argwhere(timeVecs[iCam]==tEnd))
-        keypointsSync.append(key[:,iStart:iEnd+1,:])
-        confidenceSync.append(confidence[:,iStart:iEnd+1])
-        if shiftVals[iCam] > 0:
-            shiftednNansInOut = nansInOutList[iCam] - shiftVals[iCam]
-        else:
-            shiftednNansInOut = nansInOutList[iCam]
-        nansInOutSync.append(shiftednNansInOut)        
-        # Save start and end frames to list, so can rewrite videos in
-        # triangulateMultiviewVideo
-        startEndFrames.append([iStart,iEnd])
-        
-    # Plot synchronized velocity curves
-    if visualize:
-        # Vert Velocity
-        f, (ax0,ax1) = plt.subplots(1,2)
-        for (timeVec,vertVel) in zip(timeVecs,vertVelList):
-            ax0.plot(timeVec[range(len(vertVel))],vertVel)
-        legNames = [c_cameras2Use[iCam] for iCam in range(len(vertVelList))]
-        ax0.legend(legNames)
-        ax0.set_title('summed vertical velocities')
-        
-        # Marker speed
-        for (timeVec,mkrSpeed) in zip(timeVecs,mkrSpeedList):
-            ax1.plot(timeVec[range(len(vertVel))],mkrSpeed[2])
-        legNames = [c_cameras2Use[iCam] for iCam in range(len(vertVelList))]
-        ax1.legend(legNames)
-        ax1.set_title('Right Ankle Speed')
-    
-    # Plot a single marker trajectory to see effect of filtering and occlusion removal
-    if visualize2Dkeypoint:
-        nCams = len(keypointListFilt)
-        nCols = int(np.ceil(nCams/2))
-        mkr = mkrDict['RBigToe']
-        mkrXY = 0 # x=0, y=1
-        
-        fig = plt.figure()
-        fig.set_size_inches(12,7,forward=True)
-        for camNum in range(nCams):
-            ax = plt.subplot(2,nCols,camNum+1)
-            ax.set_title(c_cameras2Use[iCam])
-            ax.set_ylabel('yPos (pixel)')
-            ax.set_xlabel('frame')
-            ax.plot(keypointListUnfilt[camNum][mkr,:,mkrXY],linewidth = 2)
-            ax.plot(keypointListOcclusionRemoved[camNum][mkr,:,mkrXY],linewidth=1.6)
-            ax.plot(keypointListFilt[camNum][mkr,:,mkrXY],linewidth=1.3)
-            
-            # find indices where conf> thresh or nan (the marker is used in triangulation)
-            usedInds = np.argwhere(np.logical_or(confidenceListFilt[camNum][mkr,:] > confidenceThreshold ,
-                                                 np.isnan(confidenceListFilt[camNum][mkr,:])))
-            if len(usedInds) > 0:
-                ax.plot(usedInds,keypointListFilt[camNum][mkr,usedInds,mkrXY],linewidth=1)
-                ax.set_ylim((.9*np.min(keypointListFilt[camNum][mkr,usedInds,mkrXY]),1.1*np.max(keypointListFilt[camNum][mkr,usedInds,mkrXY])))
-            else:
-                ax.text(0.5,0.5,'no data used',horizontalalignment='center',transform=ax.transAxes)
-            
-        plt.tight_layout()
-        ax.legend(['unfilt','occlusionRemoved','filt','used for triang'],bbox_to_anchor=(1.05,1))
-        
-    # We need to add back the cameras that have been kicked out.
-    # We just add back zeros, they will be kicked out of the triangulation.
-    idxCameras2NotUse = [cameras2Use.index(cam) for cam in cameras2NotUse]
-    for idxCamera2NotUse in idxCameras2NotUse:
-        keypointsSync.insert(idxCamera2NotUse, np.zeros(keypointsSync[0].shape))
-        confidenceSync.insert(idxCamera2NotUse, np.zeros(confidenceSync[0].shape))
-        nansInOutSync.insert(idxCamera2NotUse, np.array([np.nan, np.nan]))
-        startEndFrames.insert(idxCamera2NotUse, None)
- 
-    return keypointsSync, confidenceSync, nansInOutSync, startEndFrames
-
-# %%
-def detectHandPunchAllVideos(handPunchPositionList,sampleFreq,punchDuration=3):
-    
-    isPunch = []
-    for pos in handPunchPositionList:
-        isPunchThisVid = []
-        relPosList = []
-        maxIndList = []
-        for iSide in range(2):
-            relPos = -np.diff(pos[(iSide, iSide+2),:],axis=0) # vertical position of wrist over shoulder
-            relPosList.append(relPos)
-
-            maxInd = np.argmax(relPos)
-            maxIndList.append(maxInd)
-            maxPos = relPos[:,maxInd]
-                        
-            isPunchThisVid.append(False)
-            if maxPos>0:
-                zeroCrossings = np.argwhere(np.diff(np.sign(np.squeeze(relPos))) != 0)
-                if any(zeroCrossings>maxInd) and any(zeroCrossings<maxInd):
-                    startLength = np.abs(np.min([zeroCrossings-maxInd]))/sampleFreq
-                    endLength = np.abs(np.min([-zeroCrossings+maxInd]))
-                    if (startLength+endLength)/sampleFreq<punchDuration:
-                        isPunchThisVid[-1] = True
-                        
-        # Ensure only one arm is punching
-        hand = None
-        isPunch.append(False)
-        if isPunchThisVid[0]:
-            if relPosList[1][:,maxIndList[0]] < 0:
-                isPunch[-1] = True
-                hand = 'r'
-        elif isPunchThisVid[1]:
-            if relPosList[0][:,maxIndList[1]] < 0:
-                isPunch[-1] = True
-                hand = 'l' 
-        
-    isTrialPunch = all(isPunch)
-    
-    return isTrialPunch, hand
-
 # %% 
 def syncHandPunch(positions,hand,maxShiftSteps=600):
     if hand == 'r':
@@ -1296,13 +921,6 @@ def syncHandPunch(positions,hand,maxShiftSteps=600):
     corr_val,lag = cross_corr(relVel[1],relVel[0],multCorrGaussianStd=maxShiftSteps,visualize=False)
     
     return corr_val, lag
-
-# %% 
-def getPositions(keypoints,indsMarkers,direction=1):
-                     
-    positions = np.max(keypoints.max(axis=1)[:,1])-keypoints[indsMarkers,:,:]
-                    
-    return positions[:,:,direction]
 
 # %% 
 def detectGait(rSpeed,lSpeed,frameRate):
@@ -1363,23 +981,6 @@ def detectFeetMoving(allMarkers,confidence,ankleInds,motionThreshold=.5):
     
     return anyFootMoving
     
-
-# %% 
-def detectGaitAllVideos(mkrSpeedList,allMarkers,confidence,ankleInds,sampleFreq):
-    isGaits = []
-    feetMoving = []
-    for c_mkrSpeed,allMkrs,conf in zip(mkrSpeedList,allMarkers,confidence):
-        isGaits.append(detectGait(c_mkrSpeed[0], c_mkrSpeed[1], sampleFreq))
-        feetMoving.append(detectFeetMoving(allMkrs,conf,ankleInds))
-    if len(isGaits) > 2:
-        true_count = sum(isGaits)
-        isGait = true_count>=len(isGaits)-1
-    else:
-        isGait = all(isGaits)
-        
-    isGait = isGait and any(feetMoving)
-        
-    return isGait
 
 # %% 
 def getLargestBoundingBox(data, bbox, confThresh=0.6):
@@ -1708,121 +1309,6 @@ def trackBoundingBox(videoPath,bbStart,allPeople,allBoxes,dataOut,frameStart = 0
         
     return dataOut
 
-# %% 
-def removeOccludedSide(key2D,confidence,mkrInds,confThresh,visualize=False):
-    
-    key2D_out = np.copy(key2D)
-    confidence_out = np.copy(confidence)
-    
-    #Parameters
-    confDif = .2 # the difference in mean confidence between R and L sided markers. If dif > confDif, must be occluded
-        
-    rMkrs = mkrInds['right']
-    lMkrs = mkrInds['left']
-    
-    # nan confidences should be turned to 0 at this point...not sure why I used a nanmean
-    rConf = np.nanmean(confidence_out[rMkrs,:],axis=0)
-    lConf = np.nanmean(confidence_out[lMkrs,:],axis=0)
-    dConf = rConf-lConf
-
-    
-    if visualize:
-        plt.figure()
-        plt.plot(rConf,color='r')
-        plt.plot(lConf,color='k')
-        plt.plot(dConf,color=(.5,.5,.5))
-        plt.legend(['rightConfidence','leftConfidence','dConf'])
-        plt.xlabel('frame')
-        plt.ylabel('mean confidence')
-    
-    
-    rOccluded = np.where(np.logical_and(np.less(dConf,confDif) , np.less(rConf,confThresh)))[0]
-    lOccluded = np.where(np.logical_and(np.greater(dConf,confDif) , np.less(lConf,confThresh)))[0]
-    
-    # if inleading and exiting values are in occluded list, don't change them
-    def zeroPad(occInds,nSamps):
-    
-        # Default: not setting confidence of any indices to 0
-        zeroInds = np.empty([0],dtype=np.int64)
-        
-        if len(occInds) == 0:
-            return occInds,zeroInds
-
-        # find gaps
-        dInds = np.diff(occInds,prepend=1)
-
-        #all first vals        
-        if occInds[0] == 0:
-            #special case - occlusion starts at beginning but there are no gaps afterwards
-            if not any(dInds>1):
-                zeroInds = np.concatenate((zeroInds, np.copy(occInds)))
-                occInds = []
-                return occInds, zeroInds
-
-            else:
-                # otherwise, remove in-leading indices
-                firstGap = np.argwhere(dInds>1)[0]
-                occInds = np.delete(occInds,np.arange(firstGap))
-                dInds = np.delete(dInds,np.arange(firstGap))
-                zeroInds = np.concatenate((zeroInds, np.arange(firstGap)))
-
-        #all end vals
-        if occInds[-1] == nSamps-1:
-            #special case - occlusion goes to end but there are no gaps beforehand
-            if not any(dInds>1):
-                zeroInds = np.concatenate((zeroInds, np.copy(occInds)))
-                occInds = []
-            else:
-                # otherwise, remove end zeros
-                lastGap = np.argwhere(dInds>1)[-1]
-                occInds_init = np.copy(occInds)
-                occInds = np.delete(occInds,np.arange(lastGap,len(occInds)))
-                zeroInds = np.concatenate((zeroInds, np.arange(occInds_init[lastGap[0]],occInds_init[-1]+1)))
-
-        return occInds, zeroInds
-
-    nSamps = len(rConf)
-    # Assumes that occlusion can't happen starting in first frame or ending in last frame. Looks for first "gap"
-    rOccluded,rZeros = zeroPad(rOccluded,nSamps)
-    lOccluded,lZeros = zeroPad(lOccluded,nSamps)    
-    
-    # Couldn't figure out how to index multidimensions...
-    # Set occlusion confidences to nan. Later, the keypoints with associated nan confidence will be cubic splines, 
-    # and the confidence will get set to something like 0.5 for weighted triangulation in utilsCameraPy3 
-    for mkr in lMkrs:
-        if np.count_nonzero(confidence_out[mkr,:]>0)>3:
-            lNan = np.intersect1d(lOccluded,np.arange(np.argwhere(confidence_out[mkr,:]>0)[1],np.argwhere(confidence_out[mkr,:]>0)[-1])) # don't add nans to 0 pad
-        else:
-            lNan = np.array([],dtype='int64')
-        key2D_out[mkr,lNan.astype(int),:] = np.nan
-        confidence_out[mkr,lNan.astype(int)] = np.nan
-        confidence_out[mkr,lZeros] = 0
-        
-    for mkr in rMkrs:
-        if np.count_nonzero(confidence_out[mkr,:]>0)>3:  
-            rNan = np.intersect1d(rOccluded,np.arange(np.argwhere(confidence_out[mkr,:]>0)[1],np.argwhere(confidence_out[mkr,:]>0)[-1])) # don't add nans to 0 pad
-        else:
-            rNan = np.array([],dtype='int64')
-        key2D_out[mkr,rNan.astype(int),:] = np.nan
-        confidence_out[mkr,rNan.astype(int)] = np.nan
-        confidence_out[mkr,rZeros] = 0
-
-    
-    if visualize:
-        # TESTING
-        rConf = np.nanmean(confidence_out[rMkrs,:],axis=0)
-        lConf = np.nanmean(confidence_out[lMkrs,:],axis=0)
-        dConf = rConf-lConf
-        plt.figure()
-        plt.plot(rConf,color='r')
-        plt.plot(lConf,color='k')
-        plt.plot(dConf,color=(.5,.5,.5))
-        plt.legend(['rightFootConfidence','leftFootConfidence','dConf'])
-        plt.xlabel('frame')
-        plt.ylabel('mean confidence')
-       
-    return key2D_out, confidence_out
-
 
 # %%
 def unpackKeypointList(keypointList):
@@ -1866,53 +1352,6 @@ def undistort2Dkeypoints(pointList2D, CameraParamList, useIntrinsicMatAsP=True):
     
     return pointList2Dundistorted
 
-# %%
-def getVertVelocity(key2D):
-    vertVel = np.diff(key2D[:,:,1],axis=1)
-    vertVelTotal = np.mean(vertVel,axis=0)
-    vertVelTotal = np.append(vertVelTotal,0) # keep length
-    if not np.max(np.abs(vertVelTotal)) == 0: # only if markers were found, otherwise div by 0
-        vertVelTotal = vertVelTotal / np.max(np.abs(vertVelTotal))
-    return vertVelTotal
-
-# %%
-# get 2D speed of specified markers
-def getMarkerSpeed(key2D,idxMkrs = [0],confidence = None, confThresh = 0.2, averageVels=False):
-    c_conf = copy.deepcopy(confidence)
-    
-    diffOrder = 1 ;
-    # 2d marker speed
-    vertVel = np.abs(np.linalg.norm(np.diff(key2D,axis=1,n=diffOrder,append=np.tile(key2D[:,[-1],:],(1,diffOrder,1))),axis=2))
-    
-    # Set velocity for low confidence to 0
-    lowConf = c_conf < confThresh
-    vertVel[lowConf] = 0
-    
-    if len(idxMkrs) >1:
-        if averageVels:
-            vertVelTotal = np.mean(vertVel[idxMkrs,:],axis=0)
-            # np.append(vertVelTotal,0)
-        else:
-            vertVelTotal = vertVel[idxMkrs,:]
-            # vertVelTotal = np.hstack((vertVelTotal,np.zeros((vertVelTotal.shape[0],1)))) # keep length
-    else:
-        vertVelTotal = vertVel[idxMkrs,:]
-        # np.append(vertVelTotal,0)
-    if not np.max(np.abs(vertVelTotal)) == 0: # only if markers were found, otherwise div by 0
-        vertVelTotal = vertVelTotal / np.max(np.abs(vertVelTotal))
-    return vertVelTotal
-
-# %%
-def smoothKeypoints(key2D,sdKernel=1):
-    key2D_out = np.copy(key2D)
-    for i in range(25):
-        for j in range(2):
-            key2D_out[i,:,j] = np.apply_along_axis(
-                lambda x: gaussian_filter1d(x, sdKernel),
-                arr = key2D_out[i,:,j],
-                axis = 0)
-    return key2D_out
-
 # %% 
 def filterKeypointsButterworth(key2D,filtFreq,sampleFreq,order=4):
     key2D_out = np.copy(key2D)
@@ -1947,131 +1386,7 @@ def filter3DPointsButterworth(points3D,filtFreq,sampleFreq,order=4):
         
     return points3D_out
 
-# %%
-def clean2Dkeypoints(key2D, confidence, confidenceThreshold=0.5, nCams=2, 
-                     linearInterp=False):
-    
-    key2D_out = np.copy(key2D)
-    confidence_out = np.copy(confidence)
-    confidence_sync_out = np.copy(confidence)
-    
-    nMkrs = key2D_out.shape[0]    
-    markerNames = getOpenPoseMarkerNames()
-    
-    # Turn all 0s into nans.
-    key2D_out[key2D_out==0] = np.nan    
-    
-    # Turn low confidence values to 0s if >2 cameras and nans otherwise.
-    for i in range(nMkrs):        
-        # If a marker has at least two frames with positive confidence,
-        # then identify frames where confidence is lower than threshold.
-        if len(np.argwhere(confidence_out[i,:]>0))>2:
-            nanInds = np.where(confidence_out[i,:] < confidenceThreshold)
-        # If a marker doesn't have two frames with positive confidence,
-        # then return empty list.    
-        else:
-            nanInds = []
-            faceMarkers = getOpenPoseFaceMarkers()[0]
-            # no warning if face marker
-            if not markerNames[i] in faceMarkers:
-                print('There were <2 frames with >0 confidence for {}'.format(
-                    markerNames[i]))
-        # Turn low confidence values to 0s if >2 cameras.
-        # Frames with confidence values of 0s are ignored during triangulation.
-        if nCams>2:
-            confidence_out[i,nanInds] = 0
-        # Turn low confidence values to nans if 2 cameras.
-        # Frames with nan confidence values are splined, and nan confidences
-        # are replaced by 0.5 during triangulation.
-        else:
-            confidence_out[i,nanInds] = np.nan            
-        # Turn inleading and exiting nans into 0s
-        idx_nonnans = ~np.isnan(confidence_out[i,:])
-        idx_nonzeros  = confidence_out[i,:] != 0            
-        idx_nonnanszeros = idx_nonnans & idx_nonzeros
-        if True in idx_nonnanszeros:
-            idx_nonnanszeros_first = np.where(idx_nonnanszeros)[0][0]
-            idx_nonnanszeros_last = np.where(idx_nonnanszeros)[0][-1]
-            confidence_out[i,:idx_nonnanszeros_first] = 0
-            confidence_out[i,idx_nonnanszeros_last:] = 0 
-        else:
-            confidence_out[i,:] = 0                    
-            
-        # Turn low confidence values to 0s for confidence_sync_out whatever
-        # the number of cameras; confidence_sync_out is used for 
-        # calculating the reprojection error, and using nans rather than 
-        # 0s might affect the outcome.
-        confidence_sync_out[i,nanInds] = 0
-        # Turn inleading and exiting nans into 0s
-        idx_nonnans = ~np.isnan(confidence_sync_out[i,:])
-        idx_nonzeros  = confidence_sync_out[i,:] != 0            
-        idx_nonnanszeros = idx_nonnans & idx_nonzeros
-        if True in idx_nonnanszeros:
-            idx_nonnanszeros_first = np.where(idx_nonnanszeros)[0][0]
-            idx_nonnanszeros_last = np.where(idx_nonnanszeros)[0][-1]
-            confidence_sync_out[i,:idx_nonnanszeros_first] = 0
-            confidence_sync_out[i,idx_nonnanszeros_last:] = 0 
-        else:
-            confidence_sync_out[i,:] = 0
-        
-        # Turn keypoint values to nan if confidence is low. Keypoints with nan
-        # will be interpolated. In cases with more than 2 cameras, this will
-        # have no impact on triangulation, since corresponding confidence is 0.
-        # But with 2 cameras, we need the keypoint 2D coordinates to
-        # be interpolated. In both cases, not relying on garbage keypoints for
-        # interpolation matters when computing keypoint speeds, which are used
-        # for synchronization.            
-        key2D_out[i,nanInds,:] = np.nan
-    
-    # Helper function.
-    def nan_helper(y):
-        return np.isnan(y), lambda z: z.nonzero()[0]
-    
-    # Interpolate keypoints with nans.
-    for i in range(nMkrs):
-        for j in range(2):
-            if np.isnan(key2D_out[i,:,j]).all(): # only nans
-                key2D_out[i,:,j] = 0
-            elif np.isnan(key2D_out[i,:,j]).any():  # partially nans
-                nans, x = nan_helper(key2D_out[i,:,j])
-                if linearInterp:
-                    # Note: with linear interpolation, values are carried over
-                    # backward and forward for inleading and exiting nans.
-                    key2D_out[i,nans,j] = np.interp(x(nans), x(~nans), 
-                                                    key2D_out[i,~nans,j]) 
-                else:
-                    # Note: with cubic interpolate, values are garbage for
-                    # inleading and exiting nans.
-                    try:
-                        key2D_out[i,:,j] = pchip_interpolate(
-                            x(~nans), key2D_out[i,~nans,j], 
-                            np.arange(key2D_out.shape[1]))    
-                    except:
-                        key2D_out[i,nans,j] = np.interp(x(nans), x(~nans), 
-                                                        key2D_out[i,~nans,j])
-                        
-    # Keep track of inleading and exiting nans when less than 3 cameras.
-    if nCams>2:
-        nans_in_out = np.array([np.nan, np.nan])        
-    else:
-        _, idxFaceMarkers = getOpenPoseFaceMarkers()       
-        nans_in = []
-        nans_out = []
-        for i in range(nMkrs):
-            if i not in idxFaceMarkers:
-                idx_nans = np.isnan(confidence_out[i,:])
-                if False in idx_nans:
-                    nans_in.append(np.where(idx_nans==False)[0][0])
-                    nans_out.append(np.where(idx_nans==False)[0][-1])
-                else:
-                    nans_in.append(np.nan)
-                    nans_out.append(np.nan)                    
-        in_max = np.max(np.asarray(nans_in))
-        out_min = np.min(np.asarray(nans_out))        
-        nans_in_out = np.array([in_max, out_min])
-    
 
-    return key2D_out, confidence_out, nans_in_out, confidence_sync_out
 
 # %%
 def cross_corr(y1, y2,multCorrGaussianStd=None,visualize=False, dataForReproj=None, frameRate=60):
@@ -2905,49 +2220,6 @@ def writeTRCfrom3DKeypoints(keypoints3D, pathOutputFile, keypointNames,
     trc_file.write(pathOutputFile)   
     
     return None
-
-# %% Find indices with high confidence that overlap between cameras.
-def findOverlap(confidenceList, markers4VertVel):    
-    c_confidenceList = copy.deepcopy(confidenceList)
-    
-    # Find overlapping indices.
-    confMean = [
-        np.mean(cMat[markers4VertVel,:],axis=0) for cMat in c_confidenceList]
-    minConfLength = np.min(np.array([len(x) for x in confMean]))
-    confArray = np.array([x[0:minConfLength] for x in confMean])
-    minConfidence = np.nanmin(confArray,axis=0)
-    confThresh = .5*np.nanmax(minConfidence)
-    overlapInds = np.asarray(np.where(minConfidence>confThresh))
-    
-    # Find longest stretch of high confidence.
-    difOverlap = np.diff(overlapInds)
-    
-    # Create array with last elements of continuous stretches
-    # (more than 2 timesteps between high confidence).
-    unConfidentInds = np.argwhere(np.squeeze(difOverlap)>4)
-    
-    # Identify and select the longest continuous stretch.
-    if unConfidentInds.shape[0] > 0:        
-        # First stretch (ie, between 0 and first unconfident index).
-        stretches = []
-        stretches.append(np.arange(0, unConfidentInds[0]+1))
-        # Middle stretches.
-        if unConfidentInds.shape[0] >1:
-            for otherStretch in range(unConfidentInds.shape[0]-1):
-                stretches.append(np.arange(unConfidentInds[otherStretch]+1,
-                                           unConfidentInds[otherStretch+1]+1))
-        # Last stretch (ie, between last unconfident index and last element).
-        stretches.append(np.arange(unConfidentInds[-1]+1,
-                                   overlapInds.shape[1]))
-        # Identify longest stretch.
-        longestStretch = stretches[
-            np.argmax([stretch.shape[0] for stretch in stretches])]
-        # Select longest stretch
-        overlapInds_clean = overlapInds[0,longestStretch]
-    else:
-        overlapInds_clean = overlapInds
-        
-    return overlapInds_clean, minConfLength
 
 #%%
 def loadPklVideo(pklPath, videoFullPath, imageBasedTracker=False, poseDetector='OpenPose',
